Always link system libraries
This changes the logic for linking system libraries from parsing the
input object and the `.symbols` files, to simply include the libraries
a link time in all cases.

The `.symbols` file logic were originally added as an optimization
would allow only the libraries needed to be considered at link time.

However, this is purely a link time optimization since DCE would take
care of removing any dead code.   Its only an optimization that doesn't
really apply with the wasm backend where linking is fast and `.a`
libraries are handled efficiently.

The `.symbols` files are still used for some optional libraries such
as libhtml, libgl and libal.  However the plan is to also remove that
logic in STRICT mode too.   Once we do that we can avoid having to
run llvm-nm on each input object file.

See #8912.